refactor(ex):Move common methods to util.go
