+++ src/ex/apiExec.go
@@ -1,11 +1,6 @@
 package ex
 
 import (
-    "net/http"
-    "fmt"
-    "api"
-    "time"
-    "strings"
 )
 
 type ApiExec interface{
@@ -13,22 +8,3 @@ type ApiExec interface{
     StopExec()
 }
 
-func getSimpleClient()  http.Client{
-    DefaultClient := http.Client{Timeout: 900* time.Millisecond}
-    return  DefaultClient
-}
-func GetStatus(a api.Api, respCheck api.ApiValidator) api.ApiStatus{
-    client := getSimpleClient()
-    req,err:= http.NewRequest(a.Method, a.Url, strings.NewReader(a.Data))
-    if err !=nil{
-        fmt.Println("unable to create httpReq for:", a.Method, a.Url, a.Data)
-        return api.STATUS_YELLOW
-    }
-    res, err := client.Do(req)
-    if err !=nil {
-        fmt.Println("error executing api:", req, "\n", err)
-        return api.STATUS_RED
-    }
-    apiStat :=  respCheck.GetStatus(*res, err)
-    return apiStat
-}
